Add scl caching to abc_new.
